# Conflicts:
#	score-repo/src/main/java/org/oagi/score/repo/api/impl/jooq/businessterm/JooqBusinessTermAssignmentWriteRepository.java

+++ score-repo/src/main/java/org/oagi/score/repo/api/impl/jooq/release/JooqReleaseReadRepository.java
@@ -4,7 +4,6 @@ import org.jooq.DSLContext;
 import org.jooq.Record;
 import org.jooq.RecordMapper;
 import org.jooq.SelectOnConditionStep;
-import org.jooq.types.ULong;
 import org.oagi.score.repo.api.base.ScoreDataAccessException;
 import org.oagi.score.repo.api.impl.jooq.JooqScoreRepository;
 import org.oagi.score.repo.api.impl.utils.StringUtils;
@@ -22,7 +21,6 @@ import java.util.Date;
 
 import static org.oagi.score.repo.api.impl.jooq.entity.Tables.*;
 import static org.oagi.score.repo.api.user.model.ScoreRole.*;
-import static org.oagi.score.repo.api.user.model.ScoreRole.ADMINISTRATOR;
 
 public class JooqReleaseReadRepository
         extends JooqScoreRepository
@@ -57,7 +55,7 @@ public class JooqReleaseReadRepository
     private RecordMapper<Record, Release> mapper() {
         return record -> {
             Release release = new Release();
-            release.setReleaseId(record.get(RELEASE.RELEASE_ID).toBigInteger());
+            release.setReleaseId(record.get(RELEASE.RELEASE_ID));
             release.setGuid(record.get(RELEASE.GUID));
             release.setReleaseNum(record.get(RELEASE.RELEASE_NUM));
             release.setReleaseNote(record.get(RELEASE.RELEASE_NOTE));
@@ -68,11 +66,11 @@ public class JooqReleaseReadRepository
             release.setCreatedBy(
                     (isCreatorAdmin) ?
                             new ScoreUser(
-                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")),
                                     record.get(APP_USER.as("creator").LOGIN_ID.as("creator_login_id")),
                                     Arrays.asList(creatorRole, ADMINISTRATOR)) :
                             new ScoreUser(
-                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")),
                                     record.get(APP_USER.as("creator").LOGIN_ID.as("creator_login_id")),
                                     creatorRole));
 
@@ -81,11 +79,11 @@ public class JooqReleaseReadRepository
             release.setLastUpdatedBy(
                     (isUpdaterAdmin) ?
                             new ScoreUser(
-                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")),
                                     record.get(APP_USER.as("updater").LOGIN_ID.as("updater_login_id")),
                                     Arrays.asList(updaterRole, ADMINISTRATOR)) :
                             new ScoreUser(
-                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")),
                                     record.get(APP_USER.as("updater").LOGIN_ID.as("updater_login_id")),
                                     updaterRole));
 
@@ -103,12 +101,12 @@ public class JooqReleaseReadRepository
         Release release;
         if (request.getReleaseId() != null) {
             release = (Release) select()
-                    .where(RELEASE.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())))
+                    .where(RELEASE.RELEASE_ID.eq(request.getReleaseId()))
                     .fetchOne(mapper());
         } else if (request.getTopLevelAsbiepId() != null) {
             release = (Release) select()
                     .join(TOP_LEVEL_ASBIEP).on(RELEASE.RELEASE_ID.eq(TOP_LEVEL_ASBIEP.RELEASE_ID))
-                    .where(TOP_LEVEL_ASBIEP.TOP_LEVEL_ASBIEP_ID.eq(ULong.valueOf(request.getTopLevelAsbiepId())))
+                    .where(TOP_LEVEL_ASBIEP.TOP_LEVEL_ASBIEP_ID.eq(request.getTopLevelAsbiepId()))
                     .fetchOne(mapper());
         } else if (StringUtils.hasLength(request.getReleaseNum())) {
             release = (Release) select()
